새로운 Checkbox로 대체했다
