+++ packages/co-design-core/src/components/Checkbox/stories/Checkbox.stories.tsx
@@ -18,6 +18,10 @@ export const Disabled = () => {
   return <Checkbox disabled />;
 };
 
+export const DisabledCheck = () => {
+  return <Checkbox disabled checked />;
+};
+
 export const Indeterminate = () => {
   return <Checkbox indeterminate />;
 };
